+++ src/mesido/pycml/component_library/milp/__init__.py
@@ -30,12 +30,13 @@ from .heat.heat_port import HeatPort
 from .heat.heat_pump import HeatPump
 from .heat.heat_source import HeatSource
 from .heat.heat_source_gas import HeatSourceGas
+from .heat.heat_source_elec import HeatSourceElec
 from .heat.heat_two_port import HeatTwoPort
 from .heat.low_temperature_ates import LowTemperatureATES
 from .heat.node import Node
 from .heat.pump import Pump
 from .multicommodity.airwater_heat_pump_elec import AirWaterHeatPumpElec
-from .multicommodity.electro_boiler import ElecBoiler
+from .multicommodity.elec_heat_source_elec import ElecHeatSourceElec
 from .multicommodity.electrolyzer import Electrolyzer
 from .multicommodity.gas_heat_source_gas import GasHeatSourceGas
 
@@ -50,7 +51,8 @@ __all__ = [
     "Compressor",
     "ControlValve",
     "HeatDemand",
-    "ElecBoiler",
+    "HeatSourceElec",
+    "ElecHeatSourceElec",
     "ElectricityCable",
     "ElectricityDemand",
     "ElectricityNode",