+++ src/mesido/pycml/component_library/milp/heat/__init__.py
@@ -14,6 +14,7 @@ from .heat_port import HeatPort
 from .heat_pump import HeatPump
 from .heat_source import HeatSource
 from .heat_source_gas import HeatSourceGas
+from .heat_source_elec import HeatSourceElec
 from .heat_two_port import HeatTwoPort
 from .low_temperature_ates import LowTemperatureATES
 from .node import Node
@@ -37,6 +38,7 @@ __all__ = [
     "HeatTwoPort",
     "HeatPipe",
     "HeatSource",
+    "HeatSourceElec",
     "LowTemperatureATES",
     "Node",
     "Pump",