+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Adding constraint so that all heating demands are placed at the end of the simulation
 - HeatSourceGas (gas boiler with 2 ports)is supported for DTK and its cost attributes are supported in Mesido
 - Cost attributes of GasHeatSourceGas (gas boiler with 3 ports) are supported in Mesido
+- HeatSourceElec (e-boiler with 2 ports) is supported for DTK and its cost attributes are supported in Mesido
+- Cost attributes of ElecHeatSourceElec (e-boiler with 3 ports) are supported in Mesido
 
 ## Changed
 - Clean up of old code and removing duplicates.